backend/requirements.txt
fastapi == 0.101.1
uvicorn[standard]
pydantic==1.10.11
python-dotenv
email-validator
motor==2.5.1
requests
pytest
argon2-cffi
beanie
celery[redis]
python-jose[cryptography]
fastapi-jwt-auth